+++ docs/changelog.rst
@@ -18,6 +18,9 @@ New Features
 - The :code:`scaleAndOffset()` extended method for ee.Image and ee.ImageCollection classes was created and will replace the :code:`scale()` method.
 - The :code:`spectralIndices()` extended method for ee.Image and ee.ImageCollection classes was created and will replace the :code:`index()` method.
 - The :code:`preprocess()` extended method for ee.Image and ee.ImageCollection classes was created.
+- The :code:`getDOI()` extended method for ee.Image and ee.ImageCollection classes was created.
+- The :code:`getCitation()` extended method for ee.Image and ee.ImageCollection classes was created.
+- The :code:`getSTAC()` extended method for ee.Image and ee.ImageCollection classes was created.
 - The binary operators (+, -, \*\, /, //, %, \**\ , <<, >>, &, |) were overloaded for ee.Number objects.
 - The binary operators (+, \*\) were overloaded for ee.List objects.
 - The rich comparisons (<, <=, ==, !=, >, >=) were overloaded for ee.Number objects.